requirements.txt
easydict==1.7
matplotlib==3.0.3
numpy==1.16.3
numpydoc==0.8.0
opencv-python==3.4.2.16
pandas==0.23.4
Pillow==6.0.0
scikit-image==0.15.0
scikit-learn==0.20.1
scipy==1.2.1
